api/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.23.2
strawberry-graphql>=0.209.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
slowapi>=0.1.8

# Database
aiosqlite>=0.19.0

# Minecraft
mcstatus>=11.0.0

# SFTP
paramiko>=3.3.1

# Utils
python-dotenv>=1.0.0